+++ domain/src/main/java/net/boreeas/riotapi/constants/Season.java
@@ -22,10 +22,10 @@ package net.boreeas.riotapi.constants;
 public enum Season {
     PRESEASON3(3),
     SEASON3(3),
-    PRESEASON4(4),
-    SEASON4(4),
-    PREASEASON5(5),
-    SEASON5(5);
+    PRESEASON2014(4),
+    SEASON2014(4),
+    PREASEASON2015(5),
+    SEASON2015(5);
 
     public final int numeric;