+++ rest/src/main/java/net/boreeas/riotapi/rest/Match.java
@@ -20,6 +20,8 @@ import lombok.Getter;
 import net.boreeas.riotapi.com.riotgames.platform.game.GameMode;
 import net.boreeas.riotapi.com.riotgames.platform.game.GameType;
 import net.boreeas.riotapi.com.riotgames.platform.game.QueueType;
+import net.boreeas.riotapi.com.riotgames.team.dto.Team;
+import net.boreeas.riotapi.constants.Season;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -41,4 +43,8 @@ public class Match {
     private GameMode matchMode;
     private GameType matchType;
     private QueueType queueType;
+    private String platformId;
+    private List<Team> teams = new ArrayList<>();
+    private MatchTimeline timeline;
+    private Season season;
 }